improve initialization of xpredict data, create models and scripts folder
