+++ openpredict/models/openpredict_model.py
@@ -1090,439 +1090,3 @@ def get_predictions(
     #     ],
     #     "umls": "C0212166"
     #     },
-
-
-# @Author Arif YILMAZ, a.yilmaz@maastrichtuniversity.nl
-#@repoaddress "https://github.com/arifx/XPREDICT"
-
-import csv
-
-import pandas as pd
-import numpy as np
-import torch as th
-import torch.nn.functional as fn
-import os
-import sys
-
-
-
-def predictDrugRepositioning(diseaseCURIElist,noofResults):
-  FILEPATH="/openpredict/data/kgpredict/"
-  #FILEPATH="/data/kgpredict/"
-  EMBPATH=FILEPATH+"embed/"
-
-  try:
-    #cwd = os.getcwd() + "/translator-openpredict/drugrepositioningfilesmodels"
-    cwd = os.getcwd() 
-
-    diseaselist= [diseaseCURIElist]
-    #print("DISLIST:"+str(diseaselist))
-    drug_list = []
-
-    with open(cwd+FILEPATH+"kgpredict_drug_diseasemappings.tsv", newline='', encoding='utf-8') as csvfile:
-        reader = csv.DictReader(csvfile, delimiter='\t', fieldnames=['drug','ids'])
-        for row_val in reader:
-            drug_list.append(row_val['drug'])
-    
-    len(drug_list)
-
-    treatment = ['Hetionet::CtD::Compound:Disease','GNBR::T::Compound:Disease']
-    
-    entity_idmap_file = cwd+EMBPATH+'entities.tsv'
-    relation_idmap_file = cwd+EMBPATH+'relations.tsv'
-    
-    entity_map = {}
-    entity_id_map = {}
-    relation_map = {}
-     
-    with open(entity_idmap_file, newline='', encoding='utf-8') as csvfile:
-         
-        reader = csv.DictReader(csvfile, delimiter='\t', fieldnames=['name','id'])
-        for row_val in reader:
-            entity_map[row_val['name']] = int(row_val['id'])
-            entity_id_map[int(row_val['id'])] = row_val['name']
-          
-    with open(relation_idmap_file, newline='', encoding='utf-8') as csvfile:
-        reader = csv.DictReader(csvfile, delimiter='\t', fieldnames=['name','id'])
-        for row_val in reader:
-            relation_map[row_val['name']] = int(row_val['id'])
-
-    drug_ids = []
-    disease_ids = []
-    for drug in drug_list:
-        drug_ids.append(entity_map[drug])
-       
-    for disease in diseaselist:
-        disease_ids.append(entity_map["Disease::"+str(disease)])
-
-    treatment_rid = [relation_map[treat]  for treat in treatment]
-  
-    # Load embeddings
-    pth=cwd+EMBPATH+'entity_embeddings.npy'
-    #print("DISLIST3:"+str(pth))  
-    entity_emb = np.load(pth)
-    #np.savez_compressed("entity_embeddings.npz", entity_emb=entity_emb)
-    rel_emb = np.load(cwd+EMBPATH+'relation_embeddings.npy')
-
-         
-    drug_ids = th.tensor(drug_ids).long()
-    disease_ids = th.tensor(disease_ids).long()
-    treatment_rid = th.tensor(treatment_rid)
-    
-    drug_emb = th.tensor(entity_emb[drug_ids])
-    treatment_embs = [th.tensor(rel_emb[rid]) for rid in treatment_rid]
-
-    
-    gamma=12.0
-    def getEmbedding(head, rel, tail):
-        score = head + rel - tail
-        return gamma - th.norm(score, p=2, dim=-1)
-   
-    scores_per_disease = []
-    dids = []
-    for rid in range(len(treatment_embs)):
-        treatment_emb=treatment_embs[rid]
-        for disease_id in disease_ids:
-            disease_emb = entity_emb[disease_id]
-            score = fn.logsigmoid(getEmbedding(drug_emb, treatment_emb, disease_emb))
-            scores_per_disease.append(score)
-            dids.append(drug_ids)
-    scores = th.cat(scores_per_disease)
-    dids = th.cat(dids)
-    
-    # sort scores 
-    idx = th.flip(th.argsort(scores), dims=[0])
-    scores = scores[idx].numpy()
-    dids = dids[idx].numpy()
-    
-    
-    _, unique_indices = np.unique(dids, return_index=True)
-    topk=noofResults
-    topk_indices = np.sort(unique_indices)[:topk]
-    proposed_dids = dids[topk_indices]
-    proposed_scores = scores[topk_indices]
-    
-    
-    for i in range(topk):
-        drug = int(proposed_dids[i])
-        score = proposed_scores[i]
-      
-    proposed_drugnames={}
-    
-    for row_val in range(topk):
-        drug = int(proposed_dids[row_val])
-        proposed_drugnames[row_val] = entity_id_map[drug].replace("Compound::","")
-
-    
-    prediction_df = pd.DataFrame(list(zip(
-        proposed_drugnames.values(), proposed_drugnames.values(), list(proposed_scores))), columns=['drug', 'disease', 'score'])
-    prediction_df.sort_values(by='score', inplace=True, ascending=False)
-    # prediction_df = pd.DataFrame( list(zip(pairs[:,0], pairs[:,1], y_proba[:,1])), columns =[drug_column_label,disease_column_label,'score'])
-
-    # Add namespace to get CURIEs from IDs
-    prediction_df["drug"] = "DRUGBANK:" + prediction_df["drug"]
-    prediction_df["disease"] =  prediction_df["disease"]
-    
-
-    # prediction_results=prediction_df.to_json(orient='records')
-    prediction_results = prediction_df.to_dict(orient='records')
-    return prediction_results
-  except Exception as e:
-     print(e)
-     
-  return pd.DataFrame()
-
-
-
-   
-#print("Drug Repositioning  TOP 100 Predictions (Score=0 is best drug) :")
-#drugscores=predictDrugRepositioning(uterinecancerdisease,noofResults=100)    
-#print(drugscores)
-    
-    
-     
-
-
-
-def get_drugrepositioning_results(
-        diseaseCURIElist, n_results=100, 
-    ):
-    """Run kg evaluation to get predictions
-
-    :param score: score minimum of predictions
-    :param n_results: number of predictions to return
-    :return: predictions in array of JSON object
-    """
-    # classifier: Predict OMIM-DrugBank
-    # TODO: improve when we will have more classifier
-    id_to_predict=diseaseCURIElist
-    predictions_array = predictDrugRepositioning(diseaseCURIElist,n_results)
-    print(predictions_array)
-   
-    if n_results:
-        # Predictions are already sorted from higher score to lower
-        predictions_df = predictions_array[:n_results]
-    predictions_df    
-    # Build lists of unique node IDs to retrieve label
-    predicted_ids = set([])
-    for prediction in predictions_array:
-        for key, value in prediction.items():
-            if key != 'score':
-                predicted_ids.add(value)
-    labels_dict = get_entities_labels(predicted_ids)
-
-    # TODO: format using a model similar to BioThings:
-    # cf. at the end of this file
-
-    # Add label for each ID, and reformat the dict using source/target
-    labelled_predictions = []
-    # Second array with source and target info for the reasoner query resolution
-    source_target_predictions = []
-    for prediction in predictions_array:
-        labelled_prediction = {}
-        source_target_prediction = {}
-        for key, value in prediction.items():
-            if key == 'score':
-                labelled_prediction['score'] = value
-                source_target_prediction['score'] = value
-            elif value == id_to_predict:
-                # Only for the source_target_prediction object
-                #print("SHAPDisease:"+value)
-                # for finding disease explanalations: shaps=str(xp.getXPREDICTExplanation(drugId=value))
-                source_target_prediction['source'] = {
-                    'id': id_to_predict,
-                    'type': key
-                }
-                try:
-                    if id_to_predict in labels_dict and labels_dict[id_to_predict] and labels_dict[id_to_predict]['id'] and labels_dict[value]['id']['label']:
-                        source_target_prediction['source']['label'] = labels_dict[id_to_predict]['id']['label']
-                except:
-                    print('No label found for ' + value)
-            else:
-                labelled_prediction['id'] = value
-                labelled_prediction['type'] = key
-                #print("SHAPX:"+value)
-                #SHAPDISABLE shaps=xp.getXPREDICTExplanation(drugId=value)
-             
-                #SHAPDISABLE labelled_prediction['shap'] = shaps
-                # Same for source_target object
-                source_target_prediction['target'] = {
-                    'id': value,
-                    'type': key,
-                    #SHAPDISABLE shap' : shaps
-                }
-                try:
-                    if value in labels_dict and labels_dict[value]:
-                        labelled_prediction['label'] = labels_dict[value]['id']['label']
-                        source_target_prediction['target']['label'] = labels_dict[value]['id']['label']
-                except:
-                    print('No label found for ' + value)
-                # if value in labels_dict and labels_dict[value] and labels_dict[value]['id'] and labels_dict[value]['id']['label']:
-                #     labelled_prediction['label'] = labels_dict[value]['id']['label']
-                #     source_target_prediction['target']['label'] = labels_dict[value]['id']['label']
-
-        labelled_predictions.append(labelled_prediction)
-        source_target_predictions.append(source_target_prediction)
-        # returns
-        # { score: 12,
-        #  source: {
-        #      id: DB0001
-        #      type: drug,
-        #      label: a drug
-        #  },
-        #  target { .... }}
-    #print(source_target_predictions)
-    #print(labelled_predictions)
-    
-    
-    return labelled_predictions,source_target_predictions
-
-#XPREDICT framework may be used to add explanation featuers to drug repositioning applications 
-
-
-
-import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-from sklearn.preprocessing import normalize
-from sklearn.model_selection import train_test_split
-from subprocess import check_output
-import shap as shap
-import os
-import sklearn
-import json
-
-def getSHAPModel(datasetFile="/openpredict/data/deepdrug_repurposingpredictiondataset.csv",mlmodel="LR"):
-    cwd = os.getcwd()   
-    features= ['GO-SIM_HPO-SIM',
-     'GO-SIM_PHENO-SIM',
-     'PPI-SIM_HPO-SIM',
-     'PPI-SIM_PHENO-SIM',
-     'SE-SIM_HPO-SIM',
-     'SE-SIM_PHENO-SIM',
-     'TARGETSEQ-SIM_HPO-SIM',
-     'TARGETSEQ-SIM_PHENO-SIM',
-     'TC_HPO-SIM',
-     'TC_PHENO-SIM']
-    outclass=['Class']
-    
-
-    deepdrug_dataset_df=pd.read_csv(cwd+datasetFile)
-    Xdf= deepdrug_dataset_df[features]
-    ydf=deepdrug_dataset_df[outclass]
-
-    modellr = sklearn.linear_model.LogisticRegression()
-    modellr.fit(Xdf, ydf)
-    
-    
-    # explain the model's predictions using SHAP
-    # (same syntax works for LightGBM, CatBoost, scikit-learn, transformers, Spark, etc.)
-    X100 = shap.utils.sample(Xdf, 50130)
-    explainerlr = shap.Explainer(modellr,X100)
-
-    shap_valueslr = explainerlr(Xdf)
-    
-    return shap_valueslr
-
-
-def getXPREDICTExplanation(shap_values,drugId="000"):
-    
-    if drugId=="000":
-        return pd.DataFrame().to_json()
-    
-    ##shap.plots.bar(shap_valueslr,max_display=11)
-    ##shap.plots.beeswarm(shap_valueslr)
-    # shap_valueslr[1]
-    # visualize the first prediction's explanation
-    # shap.plots.waterfall(shap_valueslr[1])
-
-    cwd = os.getcwd()   
-    #plt.show()
-    datasetFile="/openpredict/data/deepdrug_repurposingpredictiondataset.csv"
-    deepdrug_dataset_df=pd.read_csv(cwd+datasetFile)
-    
-    i=50#deepdrug_dataset_df['Drug'== drugId]
-    strx=str(drugId)
-    drid=strx[strx.rfind(":")+1:]
-    #print("DRID:"+ str( drid))
-    #print("DURGLIST:"+str(deepdrug_dataset_df.loc[deepdrug_dataset_df.Drug==drid]))
-    if deepdrug_dataset_df.shape[0] >0:
-        #shapx=deepdrug_dataset_df.loc[deepdrug_dataset_df.Drug==drid].iloc[0] #test
-        ishap=deepdrug_dataset_df.loc[deepdrug_dataset_df.Drug==drid].index[0]
-        
-        shapx=shap_values[ishap]
-        #print("ISHAP:"+str("Shapley Values:"+str(shapx)))
-        shapjson=json.dumps(str(shapx))
-        #shapjson=str(shapx)
-        #shapjson=shapx.to_json()
-        return shapjson
-    
-    else:
-        print("test ravel3")
-        return pd.DataFrame().to_json() #return null DF
-        
-    
-
-
-def get_explanations(
-        id_to_predict, model_id, app,
-        min_score=None, max_score=None, n_results=None, 
-    ):
-    """Run classifiers to get predictions
-
-    :param id_to_predict: Id of the entity to get prediction from
-    :param classifier: classifier used to get the predictions
-    :param score: score minimum of predictions
-    :param n_results: number of predictions to return
-    :return: predictions in array of JSON object
-    """
-    # classifier: Predict OMIM-DrugBank
-    # TODO: improve when we will have more classifier
-    predictions_array = query_omim_drugbank_classifier(id_to_predict, model_id)
-    shap_valueslr= getSHAPModel(datasetFile="/openpredict/data/deepdrug_repurposingpredictiondataset.csv")
-    
-    if min_score:
-        predictions_array = [
-            p for p in predictions_array if p['score'] >= min_score]
-    if max_score:
-        predictions_array = [
-            p for p in predictions_array if p['score'] <= max_score]
-    if n_results:
-        # Predictions are already sorted from higher score to lower
-        predictions_array = predictions_array[:n_results]
-
-    # Build lists of unique node IDs to retrieve label
-    predicted_ids = set([])
-    for prediction in predictions_array:
-        for key, value in prediction.items():
-            if key != 'score':
-                predicted_ids.add(value)
-    labels_dict = get_entities_labels(predicted_ids)
-
-    # TODO: format using a model similar to BioThings:
-    # cf. at the end of this file
-
-    # Add label for each ID, and reformat the dict using source/target
-    labelled_predictions = []
-    # Second array with source and target info for the reasoner query resolution
-    source_target_predictions = []
-    for prediction in predictions_array:
-        labelled_prediction = {}
-        source_target_prediction = {}
-        for key, value in prediction.items():
-            if key == 'score':
-                labelled_prediction['score'] = value
-                source_target_prediction['score'] = value
-            elif value == id_to_predict:
-                # Only for the source_target_prediction object
-                #print("SHAPDisease:"+value)
-                # for finding disease explanalations: shaps=str(xp.getXPREDICTExplanation(drugId=value))
-                source_target_prediction['source'] = {
-                    'id': id_to_predict,
-                    'type': key
-                }
-                try:
-                    if id_to_predict in labels_dict and labels_dict[id_to_predict] and labels_dict[id_to_predict]['id'] and labels_dict[value]['id']['label']:
-                        source_target_prediction['source']['label'] = labels_dict[id_to_predict]['id']['label']
-                except:
-                    print('No label found for ' + value)
-            else:
-                labelled_prediction['id'] = value
-                labelled_prediction['type'] = key
-                #print("SHAPX:"+value)
-
-                shaps=getXPREDICTExplanation(shap_values=shap_valueslr, drugId=value)
-
-                labelled_prediction['shap'] = shaps
-                # Same for source_target object
-                source_target_prediction['target'] = {
-                    'id': value,
-                    'type': key,
-                    'shap' : shaps
-                }
-                try:
-                    if value in labels_dict and labels_dict[value]:
-                        labelled_prediction['label'] = labels_dict[value]['id']['label']
-                        source_target_prediction['target']['label'] = labels_dict[value]['id']['label']
-                except:
-                    print('No label found for ' + value)
-                # if value in labels_dict and labels_dict[value] and labels_dict[value]['id'] and labels_dict[value]['id']['label']:
-                #     labelled_prediction['label'] = labels_dict[value]['id']['label']
-                #     source_target_prediction['target']['label'] = labels_dict[value]['id']['label']
-
-        labelled_predictions.append(labelled_prediction)
-        source_target_predictions.append(source_target_prediction)
-        # returns
-        # { score: 12,
-        #  source: {
-        #      id: DB0001
-        #      type: drug,
-        #      label: a drug
-        #  },
-        #  target { .... }}
-    #print(source_target_predictions)
-    #print(labelled_predictions)
-    
-    
-    return labelled_predictions,source_target_predictions
-
